caching no longer used (code still there but commented out), made performance very bad for all output media progress bar for writing is now in % mode, seems to work cancelling writing really does work now
git-svn-id: svn://littlesvr.ca/prog/isomaster/src/bk@244 ea4051ab-32fe-0310-aa8c-8a8c8f7b8cea

+++ bkGet.c
@@ -25,7 +25,7 @@
 * Public function
 * Estimates the size of the directory trees + file contents on the iso
 * */
-unsigned bk_estimate_iso_size(const VolInfo* volInfo, int filenameTypes)
+off_t bk_estimate_iso_size(const VolInfo* volInfo, int filenameTypes)
 {
     //!! add path tables, boot record
     return estimateIsoSize(&(volInfo->dirTree), filenameTypes);
@@ -76,10 +76,10 @@ const char* bk_get_volume_name(const VolInfo* volInfo)
 * Recursive
 * Estimate the size of the directory trees + file contents on the iso
 * */
-unsigned estimateIsoSize(const BkDir* tree, int filenameTypes)
+off_t estimateIsoSize(const BkDir* tree, int filenameTypes)
 {
-    unsigned estimateDrSize;
-    unsigned thisDirSize;
+    off_t estimateDrSize;
+    off_t thisDirSize;
     int numItems; /* files and directories */
     BkFileBase* child;
