+++ bkGet.h
@@ -1,3 +1,3 @@
-unsigned estimateIsoSize(const BkDir* tree, int filenameTypes);
+off_t estimateIsoSize(const BkDir* tree, int filenameTypes);
 int getDirFromString(const BkDir* tree, const char* pathStr, 
                      BkDir** dirFoundPtr);
